+++ overlays/proxy/src/main/java/com/velocitypowered/proxy/VelocityServer.java
@@ -292,6 +292,7 @@ public class VelocityServer implements ProxyServer, ForwardingAudience {
     this.doStartupConfigLoad();
 
     Conduit.init(Path.of("."));
+    Conduit.get().installBundledSpark();
 
     registerTranslations();